LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/iframe-synchronously-discard-expected.txt


FAIL IFrame discards are processed synchronously assert_equals: child window should be discarded expected null but got object "[object Window]"
